backend/requirements.txt
# Core API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# PDF Extraction
PyMuPDF>=1.23.8

# DOCX Extraction
python-docx>=1.0.0

# HTTP Client (async)
httpx>=0.25.0

# File handling
python-multipart>=0.0.6

# PDF Generation
weasyprint>=60.0

# Utility
python-dotenv>=1.0.0

# CORS
# (included in fastapi)

# Dev dependencies
pytest>=7.4.0
pytest-asyncio>=0.23.0
